extract integrations start modal to WelcomeIntegrationsPage
